Move C/C++ raw to its own header
That way users don't see them without explicitly including it
