+++ cscore/src/main/native/include/cscore_c.h
@@ -110,15 +110,6 @@ typedef struct CS_VideoMode {
   int fps;
 } CS_VideoMode;
 
-typedef struct CS_RawFrame {
-  char* data;
-  int dataLength;
-  int pixelFormat;
-  int width;
-  int height;
-  int totalData;
-} CS_RawFrame;
-
 /**
  * Property kinds
  */
@@ -283,9 +274,6 @@ CS_Source CS_CreateHttpCameraMulti(const char* name, const char** urls,
                                    CS_Status* status);
 CS_Source CS_CreateCvSource(const char* name, const CS_VideoMode* mode,
                             CS_Status* status);
-
-CS_Source CS_CreateRawSource(const char* name, const CS_VideoMode* mode,
-                             CS_Status* status);
 /** @} */
 
 /**
@@ -371,8 +359,6 @@ char** CS_GetHttpCameraUrls(CS_Source source, int* count, CS_Status* status);
  */
 void CS_PutSourceFrame(CS_Source source, struct CvMat* image,
                        CS_Status* status);
-void CS_PutRawSourceFrame(CS_Source source, const struct CS_RawFrame* image,
-                          CS_Status* status);
 void CS_NotifySourceError(CS_Source source, const char* msg, CS_Status* status);
 void CS_SetSourceConnected(CS_Source source, CS_Bool connected,
                            CS_Status* status);
@@ -397,13 +383,6 @@ CS_Sink CS_CreateCvSink(const char* name, CS_Status* status);
 CS_Sink CS_CreateCvSinkCallback(const char* name, void* data,
                                 void (*processFrame)(void* data, uint64_t time),
                                 CS_Status* status);
-
-CS_Sink CS_CreateRawSink(const char* name, CS_Status* status);
-
-CS_Sink CS_CreateRawSinkCallback(const char* name, void* data,
-                                 void (*processFrame)(void* data,
-                                                      uint64_t time),
-                                 CS_Status* status);
 /** @} */
 
 /**
@@ -445,10 +424,6 @@ void CS_SetSinkDescription(CS_Sink sink, const char* description,
 uint64_t CS_GrabSinkFrame(CS_Sink sink, struct CvMat* image, CS_Status* status);
 uint64_t CS_GrabSinkFrameTimeout(CS_Sink sink, struct CvMat* image,
                                  double timeout, CS_Status* status);
-uint64_t CS_GrabRawSinkFrame(CS_Sink sink, struct CS_RawFrame* rawImage,
-                             CS_Status* status);
-uint64_t CS_GrabRawSinkFrameTimeout(CS_Sink sink, struct CS_RawFrame* rawImage,
-                                    double timeout, CS_Status* status);
 char* CS_GetSinkError(CS_Sink sink, CS_Status* status);
 void CS_SetSinkEnabled(CS_Sink sink, CS_Bool enabled, CS_Status* status);
 /** @} */
@@ -501,9 +476,6 @@ void CS_Shutdown(void);
  * @defgroup cscore_utility_cfunc Utility Functions
  * @{
  */
-
-void CS_AllocateRawFrameData(CS_RawFrame* frame, int requestedSize);
-void CS_FreeRawFrameData(CS_RawFrame* frame);
 
 CS_UsbCameraInfo* CS_EnumerateUsbCameras(int* count, CS_Status* status);
 void CS_FreeEnumeratedUsbCameras(CS_UsbCameraInfo* cameras, int count);
