+++ cscore/src/main/native/include/cscore_cpp.h
@@ -56,19 +56,6 @@ struct UsbCameraInfo {
   std::vector<std::string> otherPaths;
 };
 
-struct RawFrame : public CS_RawFrame {
-  RawFrame() {
-    data = nullptr;
-    dataLength = 0;
-    pixelFormat = CS_PIXFMT_UNKNOWN;
-    width = 0;
-    height = 0;
-    totalData = 0;
-  }
-
-  ~RawFrame() { CS_FreeRawFrameData(this); }
-};
-
 /**
  * Video mode
  */
@@ -214,8 +201,6 @@ CS_Source CreateHttpCamera(const wpi::Twine& name,
                            CS_HttpCameraKind kind, CS_Status* status);
 CS_Source CreateCvSource(const wpi::Twine& name, const VideoMode& mode,
                          CS_Status* status);
-CS_Source CreateRawSource(const wpi::Twine& name, const VideoMode& mode,
-                          CS_Status* status);
 /** @} */
 
 /**
@@ -302,8 +287,6 @@ std::vector<std::string> GetHttpCameraUrls(CS_Source source, CS_Status* status);
  * @{
  */
 void PutSourceFrame(CS_Source source, cv::Mat& image, CS_Status* status);
-void PutSourceFrame(CS_Source source, const CS_RawFrame& image,
-                    CS_Status* status);
 void NotifySourceError(CS_Source source, const wpi::Twine& msg,
                        CS_Status* status);
 void SetSourceConnected(CS_Source source, bool connected, CS_Status* status);
@@ -330,10 +313,6 @@ CS_Sink CreateCvSinkCallback(const wpi::Twine& name,
                              std::function<void(uint64_t time)> processFrame,
                              CS_Status* status);
 
-CS_Sink CreateRawSink(const wpi::Twine& name, CS_Status* status);
-CS_Sink CreateRawSinkCallback(const wpi::Twine& name,
-                              std::function<void(uint64_t time)> processFrame,
-                              CS_Status* status);
 /** @} */
 
 /**
@@ -380,9 +359,6 @@ void SetSinkDescription(CS_Sink sink, const wpi::Twine& description,
                         CS_Status* status);
 uint64_t GrabSinkFrame(CS_Sink sink, cv::Mat& image, CS_Status* status);
 uint64_t GrabSinkFrameTimeout(CS_Sink sink, cv::Mat& image, double timeout,
-                              CS_Status* status);
-uint64_t GrabSinkFrame(CS_Sink sink, CS_RawFrame& image, CS_Status* status);
-uint64_t GrabSinkFrameTimeout(CS_Sink sink, CS_RawFrame& image, double timeout,
                               CS_Status* status);
 std::string GetSinkError(CS_Sink sink, CS_Status* status);
 wpi::StringRef GetSinkError(CS_Sink sink, wpi::SmallVectorImpl<char>& buf,
